Moved things around, made files
blackjack.R is the same as before
created a new file blackjackSimulator that has the same functionality, but uses different files for parts of the code (dealer, player,shuffle,winner etc).
There are 2 winner files, each do the same thing. One is verbose, so has the print statements (for debugging), and the other does not